requirements.txt
<<<<<<< HEAD
aiohappyeyeballs==2.4.3
aiohttp==3.11.7
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
beautifulsoup4==4.12.3
certifi==2024.8.30
click==8.1.7
deprecation==2.1.0
fastapi==0.109.2
frozenlist==1.5.0
gcloud==0.18.3
googleapis-common-protos==1.66.0
gotrue==2.10.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.26.0
hyperframe==6.0.1
idna==3.10
multidict==6.1.0
oauth2client==4.1.3
packaging==24.2
postgrest==0.18.0
propcache==0.2.0
protobuf==5.29.0rc3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.5.2
pydantic_core==2.14.5
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
realtime==2.0.6
redis==5.0.1
rsa==4.9
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.36.3
storage3==0.9.0
=======
fastapi==0.115.5
httpx==0.27.2
mangum==0.19.0
pydantic==2.10.2
python-dotenv==1.0.1
uvicorn==0.32.1
>>>>>>> origin
supabase==2.10.0
